+++ scripts/build-deb.sh
@@ -56,7 +56,7 @@ Terminal=false
 Type=Application
 Categories=Utility;Video;
 StartupWMClass=scrcpy
-Actions=Status;Config;Setup;Stop;Logs;
+Actions=Status;Config;Setup;Stop;Logs;Update;
 
 [Desktop Action Status]
 Name=Check Status
@@ -85,6 +85,12 @@ Terminal=false
 [Desktop Action Logs]
 Name=Show Logs
 Exec=/usr/bin/android-webcam-run-in-terminal logs
+Path=/usr/bin
+Terminal=false
+
+[Desktop Action Update]
+Name=Update
+Exec=/usr/bin/android-webcam-run-in-terminal update
 Path=/usr/bin
 Terminal=false
 EOF